rename to python_seafile
