+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = -s -v --doctest-modules --ignore=build --ignore=dist --ignore=seafileapi.egg-info --ignore=setup.py
+addopts = -s -v --doctest-modules --ignore=build --ignore=dist --ignore=python_seafile.egg-info --ignore=setup.py
