+++ python_seafile/__init__.py
@@ -1,5 +1,5 @@
-from seafileapi.client import SeafileApiClient
-from seafileapi._version import __version__
+from python_seafile.client import SeafileApiClient
+from python_seafile._version import __version__
 
 def connect(server, username, password):
     client = SeafileApiClient(server, username, password)
