+++ python_seafile/exceptions.py
