+++ python_seafile/repos.py
@@ -1,5 +1,5 @@
-from seafileapi.repo import Repo
-from seafileapi.utils import raise_does_not_exist
+from python_seafile.repo import Repo
+from python_seafile.utils import raise_does_not_exist
 
 class Repos(object):
     def __init__(self, client):
